fix login component
