+++ src/util/index.js
@@ -1,6 +1,6 @@
 import axios from 'axios';
 
-const BASE_URL = 'https://localhost:3000'
+const BASE_URL = '/api'
 
 export const setLocalToken = (token) => {
     localStorage.setItem("token", token);
@@ -33,24 +33,3 @@ export const removeUser = () => {
   localStorage.removeItem("user");
 };
 
-export const fetchUser = async ({method, body, url}) => {
-  try {
-    const options = {
-      method: method || 'get',
-      data: body,
-      url: `${BASE_URL}${url}`
-    }
-    console.log('url', options.url);
-
-    if (getLocalToken()) {
-      options.headers = {
-        'Authorization': `Bearer ${getLocalToken()}`
-      } 
-    }
-
-    const {data} = await axios(options);  
-    return data;
-  } catch (error) {
-    console.error(error)
-  }
-}
